# docs/swagger.yaml
basePath: /
definitions:
  ds.AppStatus:
    properties:
      id:
        type: integer
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: bitop@spatecon.ru
    name: API Support
    url: https://vk.com/bmstu_schedule
  description: Starship's flights
  license:
    name: AS IS (NO WARRANTY)
  title: SpaceY
  version: "1.0"
paths:
...
  /api/application/{id}:
    delete:
      consumes:
      - application/json
      description: Изменение статуса заявки на удален
      parameters:
      - default: 1
        description: ID заявки
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Меняем статус заявки на удален
      tags:
      - Заявки
    get:
      consumes:
      - application/json
      description: Получение заявки с услугами
      parameters:
      - default: 1
        description: ID заявки
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Одна заявка
      tags:
      - Заявки
  /api/application/admin:
    put:
      consumes:
      - application/json
      description: Изменение статуса заявки на принят или отклонен
      parameters:
      - description: Данные для обновления статуса заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.AppStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Меняем статус заявки на принят или отклонен
      tags:
      - Заявки
  /api/application/client:
    put:
      consumes:
      - application/json
      description: Изменение статуса заявки на сформирован
      parameters:
      - description: Данные для обновления статуса заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.AppStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Меняем статус заявки на сформирован
      tags:
      - Заявки
  /api/applications:
    get:
      consumes:
      - application/json
      description: Получение списка заявок
      parameters:
      - description: Фильтрация по статусу
        in: query
        name: status
        type: string
      - description: Фильтрация по дате начала
        in: query
        name: date
        type: string
      - description: Фильтрация по дате конца
        in: query
        name: date_end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неавторизованый пользователь
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Список заявок
      tags:
      - Заявки
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"